\name{engine}
\alias{engine}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Engine wear versus size data}
\description{ Data on engine wear against engine size for 19 Volvo car engines.}
\usage{
data(engine)
}
%- maybe also `usage' for other objects documented here.
\format{
A data frame with 2 columns and 19 rows. Each row refers to one engine model.
The columns are:
\describe{
\item{wear}{an index of engine wear rate.}
\item{size}{cylinder capacity in litres.}
}}

\details{ See the source for further details.
}
 
\source{
Originally from... 
\code{http://www3.bc.sympatico.ca/Volvo_Books/engine3.html}
}

\keyword{data}










