\name{sperm.comp1}
\alias{sperm.comp1}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Sperm competition data I}
\description{Data relating sperm count to time since last inter-pair
copulation and proportion of that time spent together for 15 couples living in
Manchester UK.}
\usage{
data(sperm.comp1)
}
%- maybe also `usage' for other objects documented here.
\format{
A data frame with 4 columns and 15 rows. The columns are:
\itemize{
 \item{subject}{ An identifier for the subject/couple.}
\item{time.ipc}{ Time since last inter-pair copulation, in hours.}
\item{prop.partner}{ Proportion of \code{time.ipc} that the couple had spent
together.}
\item{count}{Sperm count in millions.}
}}

\details{The sperm counts reported are total counts in ejaculate from a single
copulation, for each of 15 couples. Also recorded are the time since the
couple's previous copulation, and the proportion of that time that the couple
had spent together. The data are from volunteers from Manchester University
and were gathered to test theories about human sperm competition.
See the source article for further details.
}
 
\source{
Baker, RR and Bellis M.A. (1993) 
`Human sperm competition: ejaculate adjustment by males and the function
of masturbation'.
Animal behaviour 46:861-885
}

\keyword{data}










