\name{stomata}
\alias{stomata}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Stomatal area and CO2 }
\description{Fake data on average stomatal area for 6 trees grown under 
one of two CO2 concentrations}
\usage{
data(stomata)
}
%- maybe also `usage' for other objects documented here.
\format{
A data frame with 3 columns and 24 rows. The columns are:
\describe{
 \item{area}{mean stomatal area.}
\item{CO2}{label for which CO2 treatment the measurement relates to.}
\item{tree}{label for individual tree.}
}}

\details{The context for these simulated data is given in section 6.1 of 
the source book.
}
 
\source{The reference.}

\references{
Wood, S.N. (2006, 2017) Generalized Additive Models: An Introduction with R. CRC

}

\keyword{data}










